requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
httpx
pyotp
cryptography
pydantic-settings
neo_api_client @ git+https://github.com/Kotak-Neo/Kotak-neo-api-v2.git@v2.0.1
yfinance

pandas
numpy 
alembic